Bug #1465099 power: supply: fix pc lint data type warning
[root cause  ] data type warning
[changes     ] data type warning
[side effects] no
[self test   ]:yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: yes
[board]: ums512-1h10
[test case]: compile ok
[reviewers   ] jinfeng.lin1
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

data type waning, int and unsigned int
can't compare with each others.

+++ drivers/power/supply/charger-manager.c
@@ -2988,9 +2988,9 @@ charger_set_property(struct power_supply *psy,
 		    cm->desc->psy_charger_stat[1]) {
 			if (cm->desc->double_ic_total_limit_current &&
 			    (thermal_val.intval >=
-			    cm->desc->double_ic_total_limit_current))
+			     (int)cm->desc->double_ic_total_limit_current))
 				thermal_val.intval =
-					cm->desc->double_ic_total_limit_current;
+					(int)cm->desc->double_ic_total_limit_current;
 			thermal_val.intval /= 2;
 		}
